class10/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.124.4",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.27",
    "uuid>=1.30",
    "uvicorn>=0.38.0",
]
